V 1.3
- Removed unnecessary activty => transfered all code to MainActivity
