--- app/src/main/res/layout/activity_trim_video.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".TrimVideoActivity"
-    android:background="@drawable/background_fade"
-    android:orientation="vertical">
-
-
-
-
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
